requirements.txt
Flask>=3.0.0
google-generativeai>=0.8.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
Pillow>=10.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
supabase>=2.0.0